vb first attempt

+++ app/controllers/boards_controller.rb
@@ -1,4 +1,6 @@
 class BoardsController < ApplicationController
+  before_action :authenticate_user!, only: [:create, :update, :destroy]
+
   def index
     matching_boards = Board.all
 
@@ -24,15 +26,14 @@ class BoardsController < ApplicationController
   end
 
   def create
-    the_board = Board.new
-    the_board.name = params.fetch("query_name")
+    the_board = current_user.boards.build(name: params[:query_name])
 
-    if the_board.valid?
-      the_board.save
-      redirect_to("/boards/#{the_board.id}", { :notice => "Board created successfully." })
+    if the_board.save
+      redirect_to "/boards/#{the_board.id}", notice: "Board created successfully."
     else
-      redirect_to("/boards", { :alert => the_board.errors.full_messages.to_sentence })
+      redirect_to "/boards", alert: the_board.errors.full_messages.to_sentence
     end
+
   end
 
   def update
@@ -50,11 +51,14 @@ class BoardsController < ApplicationController
   end
 
   def destroy
-    the_id = params.fetch("path_id")
-    the_board = Board.where({ :id => the_id }).at(0)
+    the_board = Board.find(params[:path_id])
 
-    the_board.destroy
-
-    redirect_to("/boards", { :notice => "Board deleted successfully."} )
+    if the_board.user == current_user
+      the_board.destroy
+      redirect_to "/boards", notice: "Board deleted successfully."
+    else
+      redirect_to "/boards", alert: "You are not authorized to delete this board."
+    end
   end
+
 end
